Introduce caching for tracking files.
Reason:
The tracking file manager typically in large m2e projects will be used
hundrends of times (due to hundreds of artifacts). If the projects are
refreshed the files are re-read again and again. This could amount to as
much as 43% of time spent in m2e MavenBuilder (in my scenario on Windows
platform)

Changes:
This commit introduces the read cache for tracking files assuming that
those files do not change frequently (or at all).
The cache entries are verified once per minute if the underlying file
had changed its lastModified timestamp and if so they are re-read. If
the file timestamp had not changed the entry is valid for one more
minute.

Additionally since on Windows canonicalization is expensive operation it
is also being cached.

Cached properties are defensively copied when returned.

TODOs:
 - the cache is unbounded and there is no clearing policy based on LRU
 and maximum size (for instance). However for typical usages this will
 not be a problem since number of artifacts is probably couple of
 thousands and should not pose a problem for memory.